rename move_afk -> afk_mover
